% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stats_funs.R
\name{ad_Wald_CI}
\alias{ad_Wald_CI}
\title{Calculate a confident interval using the adjusted Wald method}
\usage{
ad_Wald_CI(p = NULL, n = NULL, alpha = NULL)
}
\arguments{
\item{p}{proportion of trials that were success}

\item{n}{number of trials}

\item{alpha}{desired confident level}
}
\value{
the adjusted Wald confident interval
}
\description{
Readers interested in more details about adjusted Wald confident interval
are referred to \href{http://www.measuringux.com/AdjustedWald.htm#:~:text=The\%20formula\%20for\%20calculating\%20the\%20Adjusted\%20Wald\%20confidence,the\%20z-value\%20corresponding\%20to\%20the\%20desired\%20confidence\%20level}{Calculating a Confidence Interval for Task Completion Using the Adjusted Wald Method}.
Firstly encountered in Herzog, Walter; Boomsma, Anne; Reinecke, Sven (2007): The Model-Size Effect on Traditional and Modified Tests of Covariance Structures. In Structural Equation Modeling: A Multidisciplinary Journal 14 (3), pp. 361–390. DOI: 10.1080/10705510701301602.
}
\examples{
ad_Wald_CI(0.8, 5, 0.025)
}
