% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my functions.R
\name{cofactor}
\alias{cofactor}
\title{Compute the cofactor matrix}
\usage{
cofactor(input)
}
\arguments{
\item{input}{The input matrix, should be an numerical square matrix.}
}
\value{
The cofactor matrix
}
\description{
Compute the cofactor matrix
}
