% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stats_funs.R
\name{cor2cov}
\alias{cor2cov}
\title{Convert correlation matrix into covariance matrix with given sds}
\usage{
cor2cov(R, sds)
}
\arguments{
\item{R}{Correlation matrix}

\item{sds}{Standard deviation vector}
}
\value{
Covariance matrix
}
\description{
Convert correlation matrix into covariance matrix with given sds
}
\examples{
R <- matrix(rep(0.6, 9), 3, 3)
diag(R) <- 1
cor2cov(R, c(10, 15, 12))
}
