% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_recursive_res.R
\name{find_recursive_res}
\alias{find_recursive_res}
\title{Pattern Matching}
\usage{
find_recursive_res(
  x,
  pattern = NULL,
  scale_point = 5,
  rm_empty = TRUE,
  repeats.allowed = TRUE,
  simplify = FALSE,
  method = "grep"
)
}
\arguments{
\item{x}{a character vector where matches are sought.}

\item{pattern}{character string containing either regular expressions or permutation pattern specification, see details for more information.}

\item{scale_point}{integer, the scale point, use the Likert-5 as default.}

\item{rm_empty}{logical, indicating whether to remove the elements of resulting list with length 0.}

\item{repeats.allowed}{logical, indicating whether repeated response pattern is allowed when generating using \code{gtools::permutation}.}

\item{simplify}{logical, indicating whether results should be simplified if possible.}

\item{method}{character, indicating the method used to locate recursive response.}
}
\value{
a list of the same length as argument \code{pattern} (regular expression) or resultant permutations (permutation pattern specification), of which the element housing the indices of the elements of \code{x} that yielded a match of corresponding \code{pattern}.
}
\description{
Horizontally collapse a survey data matrix or data.frame into a character vector, and search for matches to argument \code{pattern} in each element of the resultant vector.
}
\details{
\code{Pattern} accepts either

1). regular expressions, each leads to a element containing the indices of matched element of \code{x}.

2). permutations pattern specification, a character scalar, has to be in the form of "p:n1,r1|n2,r2", start with the prefix "p:", where "n" corresponding to the number of element to draw from all possible response categories, "r" indicating the number of times that the permutations of previously drawn elements being repeated consecutively. The permutations are generated using the \code{permutations} function of the package \code{gtools}. For example, "p:2,5" means that any two of all the response categories occurs 5 straight times will be detected, the resultant permutations for \code{scale_point = 5} are "1212121212", "1313131313", "1414141414", "1515151515", "2121212121" and etc.
}
\examples{
df <- data.frame(v1 = c(1, 2, 2, 4, 1),
                 v2 = c(1, 3, 2, 5, 2),
                 v3 = c(1, 2, 2, 3, 3),
                 v4 = c(1, 4, 3, 4, 5))
df
# character
find_recursive_res(df, pattern = c("111", "222", "123"))
# regular expressions
find_recursive_res(df, pattern = c("1{3}", "2{3}", "123{1}"))
# permutation pattern specification
find_recursive_res(df, pattern = "p:1,3|3,1")
}
\author{
Yujun Li
}
