% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenient_tools.R
\name{floatToBits_yujun}
\alias{floatToBits_yujun}
\title{Convert float scalar into bits}
\usage{
floatToBits_yujun(x, nbits = 53L)
}
\arguments{
\item{x}{A float scalar.}

\item{nbits}{Number of bits to convert.}
}
\value{
Converted bits object.
}
\description{
Convert float scalar into bits
}
\examples{
floatToBits_yujun(0.1)
}
