% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my functions.R
\name{mahalanobis_yujun}
\alias{mahalanobis_yujun}
\title{Compute the Mahalanobis distance of each row vector}
\usage{
mahalanobis_yujun(x, Sigma = cov(x))
}
\arguments{
\item{x}{The original data matrix}

\item{Sigma}{The covariance matrix of x}
}
\value{
The Mahalanobis distance vector.
}
\description{
Compute the Mahalanobis distance of each row vector
}
\examples{
mahalanobis_yujun(MASS::mvrnorm(100, rep(0, 3), matrix(c(1, 0.6, 0.6, 0.6, 1, 0.6, 0.6, 0.6, 1), 3, 3)))
}
