% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my functions.R
\name{round_off5}
\alias{round_off5}
\title{Round function that mimic the rounding rule used in daily life
Round to 0 if the number at specified digit is less than 5, and to 1 otherwise}
\usage{
round_off5(x, digits = 0)
}
\arguments{
\item{x}{a numerical scalar object}

\item{digits}{number of digits to keep}
}
\value{
a rounded scalar object
}
\description{
Round function that mimic the rounding rule used in daily life
Round to 0 if the number at specified digit is less than 5, and to 1 otherwise
}
\examples{
round_off5(0.5)
}
