% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stats_funs.R
\name{cofactor}
\alias{cofactor}
\title{Compute the cofactor matrix}
\usage{
cofactor(input)
}
\arguments{
\item{input}{The input matrix, should be an numerical square matrix.}
}
\value{
The cofactor matrix
}
\description{
Compute the cofactor matrix
}
\examples{
cofactor(matrix(c(1, 2, 3, 10, 5, 0, 6, 7, 9), 3, 3))
}
