% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_functions.R
\name{factanal_fit}
\alias{factanal_fit}
\title{EFA using fit indices, a wrapper of 'stats::factanal'}
\usage{
factanal_fit(data_input, nfactors = 1, fit = TRUE, IC = FALSE, ...)
}
\arguments{
\item{data_input}{input data, should be either a matrix or dataframe}

\item{nfactors}{number of factors to be tested in EFA}

\item{fit}{if 'TRUE' (default), model fit information will be extracted}

\item{IC}{if 'TRUE' (default is 'FALSE'), information criterion will be provided, i.e. AIC, BIC, EBIC}

\item{...}{}
}
\value{
a regular factanal return list with fit indice added
}
\description{
EFA using fit indices, a wrapper of 'stats::factanal'
}
\examples{
factanal_fit(data, 1)
}
