% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stats_funs.R
\name{mahalanobis_yujun}
\alias{mahalanobis_yujun}
\title{Compute the Mahalanobis distance of each row vector}
\usage{
mahalanobis_yujun(x, Sigma = stats::cov(x))
}
\arguments{
\item{x}{The original data matrix}

\item{Sigma}{The covariance matrix of x}
}
\value{
The Mahalanobis distance vector.
}
\description{
Compute the Mahalanobis distance of each row vector
}
\examples{
require(MASS)
data <- mvrnorm(100, rep(0, 2), matrix(c(1, 0.6, 0.6, 1), 2, 2))
mahalanobis_yujun(data)
}
