% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_functions.R
\name{nfactors}
\alias{nfactors}
\title{Estimate the number of factors using 11 factor retention methods}
\usage{
nfactors(data_input)
}
\arguments{
\item{data_input}{input data, should be either a matrix or dataframe}
}
\value{
a vector, containing the number of factors suggested by 11 retention methods
}
\description{
The 11 retention methods including chi square test (Bartlett test statistics),
CFI, TLI, RMSEA, CD, EKC, CAF, HULL-CFI, HULL-RMSEA, KGC, PA.
}
\examples{
library(MASS)
nfactors(mvrnorm(200, rep(0, 6), diag(6)))
}
