% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_functions.R
\name{SRMR_factanal}
\alias{SRMR_factanal}
\title{compute SRMR using factanal output}
\usage{
SRMR_factanal(factanal_results)
}
\arguments{
\item{factanal_results}{output list object provided by 'stats::factanal'}
}
\value{
SRMR
}
\description{
compute SRMR using factanal output
}
\examples{
SRMR_factanal(stats::factanal(data, 3))
}
